requirements.txt
langchain>=0.1.0
langchain-community>=0.1.0
langchain-huggingface>=0.1.2
langchain-ollama>=0.3.2
langchain-openai>=0.3.14
mlflow>=2.14.3
langsmith>=0.0.1
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pytest>=7.4.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.25.0
faiss-cpu>=1.7.0
sentence_transformers>=4.1.0
newsapi-python>=0.2.6
langserver>=0.1.0
sse_starlette>=0.1.0